refactor: rename "ErrorList" class.
There is a "ErrorList.cs" file within the "ErrorList" directory, so the class was renamed and a namespace added to match the folder structure.
