+++ src/LibraryManager.Vsix/ErrorList/TableDataSource.cs
@@ -9,7 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.ErrorList
 {
     internal class TableDataSource : ITableDataSource
     {
